moved claimtrie_tests into claimtriebranching_tests
to take advantage of the test fixture there
