+++ src/Makefile.test.include
@@ -60,7 +60,6 @@ BITCOIN_TESTS =\
   test/merkle_tests.cpp \
   test/miner_tests.cpp \
   test/multisig_tests.cpp \
-  test/claimtrie_tests.cpp \
   test/claimtriecache_tests.cpp \
   test/claimtriebranching_tests.cpp \
   test/nameclaim_tests.cpp \
